Rename SingleNews component and move data from i18n
