+++ locales/ru.json
@@ -320,32 +320,6 @@
       "show_tags_button": "Показать по тегам",
       "search_tags_input": "Введите тэг",
       "clear_tags_button": "Очистить"
-    },
-    "news": [
-      {
-        "title": "Вкусности для бездомных",
-        "description": "Какие же настали морозные дни в нашей Казани. Особенно это чувствуют бездомные люди. Не всегда удается найти теплый ночлег. Единственным спасением становятся горячие бесплатные обеды. Добрый фудтрак и сотрудники «Закята» были встречены в сквере на Б.Шахиди с большой радостью в глазах.",
-        "date": [
-          {
-            "year": "2020",
-            "month": "11",
-            "date": "1"
-          }
-        ],
-        "author": "Майя Бурганова"
-      },
-      {
-        "title": "Назер — курбан для малоимущих",
-        "description": "В фонд «Закят» принесли мясо барана назер-курбана. Сотрудники Закята распределили его подопечным семьям фонда «Закята». Среди благополучателей — многодетные матери, малоимущие, одинокие старики. Все подопечные благодарят за мясную продукцию и воздают дуа за благотворителей. Пусть Аллах одарит их барактом и дарует Рай. Амин.",
-        "date": [
-          {
-            "year": "2020",
-            "month": "11",
-            "date": "2"
-          }
-        ],
-        "author": "Ринат Гарипов"
-      }
-    ]
+    }
   }
 }
